re-arranged index and seperated basemap and wms
